+++ src/main/java/org/ex/spring/pet/bean/Animal.java
@@ -1,4 +1,4 @@
-package org.ex.spring.pets.bean;
+package org.ex.spring.pet.bean;
 
 public interface Animal {
     String getKind();
