+++ src/main/java/org/ex/spring/pet/component/Animal.java
@@ -1,4 +1,4 @@
-package org.ex.spring.pets.component;
+package org.ex.spring.pet.component;
 
 public interface Animal {
     String getKind();
