+++ src/main/java/org/ex/spring/pet/component/Dog.java
@@ -1,21 +1,17 @@
-package org.ex.spring.pets.bean;
+package org.ex.spring.pet.component;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
+@Component
+@Scope("prototype")
 public class Dog implements Animal {
+    @Value("Charlie")
     private String name;
-
-    public Dog() {
-    }
-
-    public Dog(String name) {
-        this.name = name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
 
     @Override
     public String getKind() {
