+++ src/main/java/org/ex/spring/pet/component/SpringConfig.java
@@ -1,4 +1,4 @@
-package org.ex.spring.pets.component;
+package org.ex.spring.pet.component;
 
 
 import org.springframework.context.annotation.ComponentScan;
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 
 @Configuration
-@ComponentScan("org.ex.spring.pets.component")
+@ComponentScan("org.ex.spring.pet.component")
 @PropertySource("classpath:pets.properties")
 public class SpringConfig {
 }
